dscwow-backend/requirements.txt
aniso8601==8.1.0
appdirs==1.4.3
autopep8==1.5.4
CacheControl==0.12.6
cachetools==4.2.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-RESTful==0.3.8
google-api-core==1.23.0
google-auth==1.23.0
google-cloud==0.34.0
google-cloud-core==1.4.4
google-cloud-speech==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.15
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mypy-extensions==0.4.3
packaging==20.3
peewee==3.14.0
pep517==0.8.2
progress==1.5
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.6
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.4
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
rsa==4.6
six==1.14.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
